Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
google-generativeai
faster-whisper
torch
torchaudio
elevenlabs
supabase
python-dotenv
Pillow
numpy
librosa
soundfile
pydub
python-jose[cryptography]
passlib[bcrypt]
aiofiles
aiohttp 